feat: updated provier do be injected in a better place
